Unify SuspenseCache and SuspenseQueryCache together by making client part of the key
